+++ src/router/index.ts
@@ -1,4 +1,4 @@
-import About from '@/views/About.vue';
+import Landing from '@/views/Landing.vue';
 import Overview from '@/views/Overview.vue';
 import Quests from '@/views/Quests.vue';
 import Stats from '@/views/Stats.vue';
@@ -34,8 +34,8 @@ const preserveQueryString: NavigationGuardWithThis<undefined> = (
 const routes: Array<RouteRecordRaw> = [
   {
     path: '/',
-    name: 'about',
-    component: About,
+    name: 'landing',
+    component: Landing,
     meta: {
       title: 'near stats.gallery',
       noTitleSuffix: true,
