+++ src/views/landing/Card.vue
